requirements.txt
# WARNING! This is for Termux.

aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
Cython==3.0.10
docopt==0.6.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
maturin==1.5.1
multidict==6.0.5
num2words==0.5.13
pydantic==2.7.1
pydantic_core==2.18.2
pypi==2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.11.0
yarl==1.9.4
watchdog~=6.0.0